Extract test stubs
