+++ test/stubs.rb
@@ -1,0 +1,3 @@
+require_relative 'stubs/presets'
+require_relative 'stubs/structs'
+require_relative 'stubs/jekyll'
